fix: EDA が落ちても成功扱いになる穴と、その裏に隠れていた2件
前回の実行で EDA が例外で落ちたのに、ステップは「成功」で通り、
eda.json も更新されないまま気づけなかった。原因は3つ重なっている。

1. ワークフローが失敗を握りつぶしていた
   既定の run シェルは `bash -e` で pipefail が付かない。
   全ステップが `python3 ... 2>&1 | tee x.log` の形なので、
   スクリプトが落ちても tee の終了コード0が採用される。
   `defaults.run.shell: bash` で pipefail 付きにした。
   pipefail 下では `grep ... | head -1` が SIGPIPE で落ちうるので、
   opt-in ゲートの2箇所は明示的に無害化した。

2. EDA が bool 列で落ちていた
   uptrend_end は bool。pd.to_numeric は bool を bool のまま返すため、
   quantile で「boolean subtract は使えない」で例外になっていた。
   histogram / describe_column で float に落とす。

3. scalecat_code が100%欠測だった本当の原因
   前回「/equities/master の項目名を実測せずに書いた」と報告したが、
   これは誤りだった。実測ログでは ScaleCat は届いている:
     ['CoName','CoNameEn','Code','Date','Mkt','MktNm','Mrgn','MrgnNm',
      'S17','S17Nm','S33','S33Nm','ScaleCat']
   欠測の原因は値が "TOPIX Small 2" のような文字列で、
   pd.to_numeric(errors="coerce") が全件 NaN にしていたこと。
   encode_category を追加し、数値でない列は辞書順で安定した整数に
   割り当てる（実行ごとに変わるとモデルが再現しない）。
   どの値が来ているかは必ず出す。scalecat_code を特徴量に戻した。

あわせて train_model がラベル定義名に旧 LabelConfig を出していたのを
母集団に追随させた（エンバーゴは既に RISE_HORIZON を見ていて正しい）。
未使用の META_COLS を削除。

+++ research/eda_stats.py
@@ -40,7 +40,9 @@ def histogram(v: pd.Series, bins: int = N_BINS) -> Optional[Dict]:
     ビンごとの件数。外れ値で潰れないよう 1%〜99% で範囲を切る。
     範囲外の件数は別に持ち、描画側で「範囲外 N件」と出せるようにする。
     """
-    x = pd.to_numeric(v, errors="coerce")
+    # bool のまま quantile を取ると numpy が「boolean subtract は使えない」で落ちる。
+    # to_numeric は bool を bool のまま返すので、ここで明示的に数値にする。
+    x = pd.to_numeric(v, errors="coerce").astype("float64")
     x = x[np.isfinite(x)]
     if len(x) < 10:
         return None
@@ -61,7 +63,8 @@ def histogram(v: pd.Series, bins: int = N_BINS) -> Optional[Dict]:
 
 
 def describe_column(v: pd.Series) -> Dict:
-    x = pd.to_numeric(v, errors="coerce")
+    # histogram と同じ理由で bool を float に落とす（quantile が bool で落ちる）
+    x = pd.to_numeric(v, errors="coerce").astype("float64")
     finite = x[np.isfinite(x)]
     out = {
         "n": int(len(x)),
